commit :bomb: ajustes em sincronizacao de tempo em modo coordenado

+++ influunt-api/modules/influunt-core/app/engine/GerenciadorDeEstagiosCallback.java
@@ -7,11 +7,11 @@ import org.joda.time.DateTime;
  */
 public interface GerenciadorDeEstagiosCallback {
 
-    public void onEstagioChange(int anel, Long numeroCiclos, Long tempoDecorrido, DateTime timestamp, IntervaloGrupoSemaforico intervalos);
+    public void onEstagioChange(int anel, int numeroCiclos, Long tempoDecorrido, DateTime timestamp, IntervaloGrupoSemaforico intervalos);
 
-    public void onEstagioEnds(int anel, Long numeroCiclos, Long tempoDecorrido, DateTime timestamp, IntervaloGrupoSemaforico intervalos);
+    public void onEstagioEnds(int anel, int numeroCiclos, Long tempoDecorrido, DateTime timestamp, IntervaloGrupoSemaforico intervalos);
 
-    public void onCicloEnds(int anel, Long numeroCiclos);
+    public void onCicloEnds(int anel, int numeroCiclos);
 
     public void onTrocaDePlanoEfetiva(AgendamentoTrocaPlano agendamentoTrocaPlano);
